feat: Remove scope use from OperationsTimer.
In some cases, users may not wish for the OperationsTimer to
automatically set the logging scope to the action being timed.
For example, the action name provided may be a long name which
may be prepended to each log entry making the logs more verbose.
This change removes the scope use from the OperationsTimer leaving
it up to the user to set the scope explicitly if desired.